+ Fixed Player jittering with FixedUpdate() + Fixed Walknig speed animation

+++ Assets/Scripts/Stats.cs
@@ -16,6 +16,6 @@ public class Stats : ScriptableObject
     public int def;
     public int dex;
     [Space]
-    public int spd;
+    public float spd;
 
 }
